refactor(external_services): keep CloudWatch's rules out of the metrics interface
Addresses review on #14150.

Three comments turned out to share a root cause: vendor-specific facts had
leaked into the vendor-neutral layer, as validation, as rationale, and as a
false promise.

- `Period` no longer validates. "1, 5, 10, 20, 30, or a multiple of 60" is
  CloudWatch's rule; GCP Monitoring takes any duration from 60s and
  Prometheus has no such constraint. The provider now rejects periods it
  cannot take, alongside empty batches, over-long batches and backwards
  ranges — all in one place, before a round trip is spent.

- `Labels` no longer claims that an empty set reads "across every stream
  reporting it". That is false on CloudWatch, where a dimension combination
  *identifies* a metric: an unlabelled query selects the stream published
  with no dimensions, so a metric published per instance comes back empty
  rather than aggregated. What empty labels select is now documented where
  it is true.

- `MetricSeries::window` keeps its grid arithmetic, which does generalise
  (GCP and Prometheus both omit empty intervals), but no longer justifies
  itself with `treat_missing_data`. It also stops dividing by a
  non-positive period.

Builder methods taking `mut self` are gone; request types are plain structs
the caller fills in. `metrics_service.rs` is now docs, trait and re-exports,
with the types split into `query`, `response` and `error`.
